back-end/app/requirements.txt
Click (7.0)
Flask (1.1.1)
Flask-Cors (3.0.8)
itsdangerous (1.1.0)
Jinja2 (2.10.1)
MarkupSafe (1.1.1)
pip (8.1.1)
psycopg2-binary (2.7.6.1)
pycurl (7.43.0)
pygobject (3.20.0)
python-apt (1.1.0b1+ubuntu0.16.4.5)
setuptools (20.7.0)
six (1.12.0)
unattended-upgrades (0.1)
Werkzeug (0.16.0)
wheel (0.29.0)
